pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tmf-spec-parser"
version = "0.1.0"
description = "Extract, validate and diff TMForum Open API specs — generates tmf_data.json for tmf-map and other TMF DevKit tools"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Manoj Chavan", email = "manoj.chavan23@gmail.com" }]
keywords = [
    "tmforum", "telecom", "bss", "oss", "openapi",
    "tmf638", "tmf639", "tmf641", "spec", "parser",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Internet :: WWW/HTTP",
]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "pyyaml>=6.0",
]

[project.scripts]
tmf-spec-parser = "tmf_spec_parser.cli:main"

[project.urls]
Homepage   = "https://github.com/tmf-devkit/tmf-spec-parser"
Repository = "https://github.com/tmf-devkit/tmf-spec-parser"
Issues     = "https://github.com/tmf-devkit/tmf-spec-parser/issues"

[tool.hatch.build.targets.wheel]
packages = ["tmf_spec_parser"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
